+++ app/Http/Controllers/PostController.php
@@ -25,12 +25,14 @@ use App\Achievements\UserMade700Posts;
 use App\Achievements\UserMade800Posts;
 use App\Achievements\UserMade900Posts;
 use App\Achievements\UserMadeFirstPost;
-use App\Models\Forum;
 use App\Models\Post;
 use App\Models\Topic;
+use App\Models\User;
+use App\Notifications\NewPostTag;
 use App\Repositories\ChatRepository;
-use App\Repositories\TaggedUserRepository;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Notification;
+use Exception;
 
 /**
  * @see \Tests\Todo\Feature\Http\Controllers\PostControllerTest
@@ -40,140 +42,130 @@ class PostController extends Controller
     /**
      * PostController Constructor.
      */
-    public function __construct(private readonly TaggedUserRepository $taggedUserRepository, private readonly ChatRepository $chatRepository)
+    public function __construct(private readonly ChatRepository $chatRepository)
     {
+    }
+
+    /**
+     * Posts Index.
+     */
+    public function index(): \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+    {
+        return view('forum.post.index');
     }
 
     /**
      * Store A New Post To A Topic.
      */
-    public function reply(Request $request, int $id): \Illuminate\Http\RedirectResponse
+    public function store(Request $request): \Illuminate\Http\RedirectResponse
     {
-        $user = $request->user();
-        $topic = Topic::findOrFail($id);
-        $forum = $topic->forum;
-        $category = Forum::findOrFail($forum->id);
-
-        // The user has the right to create a post here?
-        if (! $category->getPermission()->reply_topic || ($topic->state == 'close' && ! $request->user()->group->is_modo)) {
-            return \to_route('forums.index')
-                ->withErrors(\trans('forum.reply-topic-error'));
-        }
-
-        $post = new Post();
-        $post->content = $request->input('content');
-        $post->user_id = $user->id;
-        $post->topic_id = $topic->id;
-
-        $v = \validator($post->toArray(), [
+        $request->validate([
             'content'  => 'required|min:1',
-            'user_id'  => 'required',
-            'topic_id' => 'required',
+            'topic_id' => 'required|integer',
         ]);
 
-        if ($v->fails()) {
-            return \to_route('forum_topic', ['id' => $topic->id])
-                ->withErrors($v->errors());
-        }
+        $user = $request->user();
 
-        $post->save();
-        $appurl = \config('app.url');
-        $href = \sprintf('%s/forums/topics/%s?page=%s#post-%s', $appurl, $topic->id, $post->getPageNumber(), $post->id);
-        $message = \sprintf('%s has tagged you in a forum post. You can view it [url=%s] HERE [/url]', $user->username, $href);
-        if ($this->taggedUserRepository->hasTags($request->input('content'))) {
-            if ($this->taggedUserRepository->contains($request->input('content'), '@here') && $user->group->is_modo) {
-                $users = \collect([]);
+        $topic = Topic::whereRelation('forumPermissions', [
+            ['reply_topic', '=', 1],
+            ['group_id', '=', $user->group_id]
+        ])
+            ->when(! $user->group->is_modo, fn ($query) => $query->where('state', '=', 'open'))
+            ->findOrFail($request->topic_id);
 
-                $topic->posts()->get()->each(function ($p) use ($users) {
-                    $users->push($p->user);
-                });
+        $forum = $topic->forum;
 
-                $this->taggedUserRepository->messagePostUsers(
-                    'forum',
-                    $users,
-                    $user,
-                    'Staff',
-                    $post
-                );
-            } else {
-                $this->taggedUserRepository->messageTaggedPostUsers(
-                    'forum',
-                    $request->input('content'),
-                    $user,
-                    $user->username,
-                    $post
-                );
-            }
-        }
+        $post = Post::create([
+            'content'  => $request->input('content'),
+            'user_id'  => $user->id,
+            'topic_id' => $topic->id,
+        ]);
 
-        // Save last post user data to topic table
-        $topic->last_post_user_id = $user->id;
-        $topic->last_post_user_username = $user->username;
-        // Count post in topic
-        $topic->num_post = Post::where('topic_id', '=', $topic->id)->count();
-        // Update time
-        $topic->last_reply_at = $post->created_at;
-        // Save
-        $topic->save();
-        // Count posts
-        $forum->num_post = $forum->getPostCount($forum->id);
-        // Count topics
-        $forum->num_topic = $forum->getTopicCount($forum->id);
-        // Save last post user data to the forum table
-        $forum->last_post_user_id = $user->id;
-        $forum->last_post_user_username = $user->username;
-        // Save last topic data to the forum table
-        $forum->last_topic_id = $topic->id;
-        $forum->last_topic_name = $topic->name;
-        // Save
-        $forum->save();
-        // Post To Chatbox
-        $appurl = \config('app.url');
-        $postUrl = \sprintf('%s/forums/topics/%s?page=%s#post-%s', $appurl, $topic->id, $post->getPageNumber(), $post->id);
-        $realUrl = \sprintf('/forums/topics/%s?page=%s#post-%s', $topic->id, $post->getPageNumber(), $post->id);
-        $profileUrl = \sprintf('%s/users/%s', $appurl, $user->username);
+        $topic->update([
+            'last_post_user_id'       => $user->id,
+            'last_post_user_username' => $user->username,
+            'num_post'                => $topic->posts()->count(),
+            'last_reply_at'           => $post->created_at,
+        ]);
 
-        if (\config('other.staff-forum-notify') && ($forum->id == \config('other.staff-forum-id') || $forum->parent_id == \config('other.staff-forum-id'))) {
+        $forum->update([
+            'num_post'                => $forum->posts()->count(),
+            'num_topic'               => $forum->topics()->count(),
+            'last_post_user_id'       => $user->id,
+            'last_post_user_username' => $user->username,
+            'last_topic_id'           => $topic->id,
+            'last_topic_name'         => $topic->name,
+            'updated_at'              => $post->created_at,
+        ]);
+
+        // Post To Chatbox and Notify Subscribers
+        $appUrl = config('app.url');
+        $postUrl = sprintf('%s/forums/topics/%s/posts/%s', $appUrl, $topic->id, $post->id);
+        $realUrl = sprintf('/forums/topics/%s/posts/%s', $topic->id, $post->id);
+        $profileUrl = sprintf('%s/users/%s', $appUrl, $user->username);
+
+        if (config('other.staff-forum-notify') && ($forum->id == config('other.staff-forum-id') || $forum->parent_id == config('other.staff-forum-id'))) {
             $topic->notifyStaffers($user, $topic, $post);
         } else {
-            $this->chatRepository->systemMessage(\sprintf('[url=%s]%s[/url] has left a reply on topic [url=%s]%s[/url]', $profileUrl, $user->username, $postUrl, $topic->name));
+            $this->chatRepository->systemMessage(sprintf('[url=%s]%s[/url] has left a reply on topic [url=%s]%s[/url]', $profileUrl, $user->username, $postUrl, $topic->name));
+
             // Notify All Subscribers Of New Reply
-            if ($topic->first_user_poster_id != $user->id) {
+            if ($topic->first_post_user_id != $user->id) {
                 $topic->notifyStarter($user, $topic, $post);
             }
 
             $topic->notifySubscribers($user, $topic, $post);
+
+            // Achievements
+            $user->unlock(new UserMadeFirstPost());
+            $user->addProgress(new UserMade25Posts(), 1);
+            $user->addProgress(new UserMade50Posts(), 1);
+            $user->addProgress(new UserMade100Posts(), 1);
+            $user->addProgress(new UserMade200Posts(), 1);
+            $user->addProgress(new UserMade300Posts(), 1);
+            $user->addProgress(new UserMade400Posts(), 1);
+            $user->addProgress(new UserMade500Posts(), 1);
+            $user->addProgress(new UserMade600Posts(), 1);
+            $user->addProgress(new UserMade700Posts(), 1);
+            $user->addProgress(new UserMade800Posts(), 1);
+            $user->addProgress(new UserMade900Posts(), 1);
         }
 
-        //Achievements
-        $user->unlock(new UserMadeFirstPost(), 1);
-        $user->addProgress(new UserMade25Posts(), 1);
-        $user->addProgress(new UserMade50Posts(), 1);
-        $user->addProgress(new UserMade100Posts(), 1);
-        $user->addProgress(new UserMade200Posts(), 1);
-        $user->addProgress(new UserMade300Posts(), 1);
-        $user->addProgress(new UserMade400Posts(), 1);
-        $user->addProgress(new UserMade500Posts(), 1);
-        $user->addProgress(new UserMade600Posts(), 1);
-        $user->addProgress(new UserMade700Posts(), 1);
-        $user->addProgress(new UserMade800Posts(), 1);
-        $user->addProgress(new UserMade900Posts(), 1);
+        // User Tagged Notification
+        if ($user->id !== $post->user_id) {
+            preg_match_all('/@([\w\-]+)/', (string) $post->content, $matches);
+            $users = User::whereIn('username', $matches[1])->get();
+            Notification::send($users, new NewPostTag($post));
+        }
 
-        return \redirect()->to($realUrl)
-            ->withSuccess(\trans('forum.reply-topic-success'));
+        return redirect()->to($realUrl)
+            ->withSuccess(trans('forum.reply-topic-success'));
     }
 
     /**
-     * Edit Post Form.
+     * Edit A Post.
      */
-    public function postEditForm(int $id, int $postId): \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+    public function edit(Request $request, int $id): \Illuminate\Contracts\View\Factory|\Illuminate\View\View
     {
-        $topic = Topic::findOrFail($id);
-        $forum = $topic->forum;
-        $category = $forum->getCategory();
-        $post = Post::findOrFail($postId);
+        $user = $request->user();
 
-        return \view('forum.post_edit', [
+        $post = Post::find($id);
+        $topic = $post->topic()
+            ->whereRelation('forumPermissions', [
+                ['show_forum', '=', 1],
+                ['read_topic', '=', 1],
+                ['reply_topic', '=', 1],
+                ['group_id', '=', $user->group_id],
+            ])
+            ->when(! $user->group->is_modo, fn ($query) => $query->where('state', '=', 'open'))
+            ->sole();
+
+        abort_unless($user->group->is_modo || $user->id === $post->user_id, 403);
+
+        $forum = $topic->forum;
+        $category = $forum->category;
+
+        return view('forum.post.edit', [
             'topic'    => $topic,
             'forum'    => $forum,
             'post'     => $post,
@@ -182,36 +174,97 @@ class PostController extends Controller
     }
 
     /**
-     * Edit A Post In A Topic.
+     * Update A Post.
      */
-    public function postEdit(Request $request, int $postId): \Illuminate\Http\RedirectResponse
+    public function update(Request $request, int $id): \Illuminate\Http\RedirectResponse
     {
+        $request->validate([
+            'content' => 'required|min:1',
+        ]);
+
         $user = $request->user();
-        $post = Post::findOrFail($postId);
-        $postUrl = \sprintf('forums/topics/%s?page=%s#post-%s', $post->topic->id, $post->getPageNumber(), $postId);
 
-        \abort_unless($user->group->is_modo || $user->id === $post->user_id, 403);
-        $post->content = $request->input('content');
-        $post->save();
+        $post = Post::findOrFail($id);
+        $postUrl = sprintf('forums/topics/%s/posts/%s', $post->topic->id, $id);
 
-        return \redirect()->to($postUrl)
-            ->withSuccess(\trans('forum.edit-post-success'));
+        abort_unless($user->group->is_modo || $user->id === $post->user_id, 403);
+
+        abort_unless(
+            $post->topic()
+                ->whereRelation('forumPermissions', [
+                    ['reply_topic', '=', 1],
+                    ['group_id', '=', $user->group_id],
+                ])
+                ->when(! $user->group->is_modo, fn ($query) => $query->where('state', '=', 'open'))
+                ->exists(),
+            403
+        );
+
+        $post->update([
+            'content' => $request->input('content'),
+        ]);
+
+        return redirect()->to($postUrl)
+            ->withSuccess(trans('forum.edit-post-success'));
     }
 
     /**
      * Delete A Post.
      *
-     * @throws \Exception
+     * @throws Exception
      */
-    public function postDelete(Request $request, int $postId): \Illuminate\Http\RedirectResponse
+    public function destroy(Request $request, int $id): \Illuminate\Http\RedirectResponse
     {
         $user = $request->user();
-        $post = Post::with('topic')->findOrFail($postId);
+        $post = Post::with('topic')->findOrFail($id);
 
-        \abort_unless($user->group->is_modo || $user->id === $post->user_id, 403);
+        abort_unless($user->group->is_modo || $user->id === $post->user_id, 403);
+
+        $topic = $post->topic()->whereRelation('forumPermissions', [
+            ['reply_topic', '=', 1],
+            ['group_id', '=', $user->group_id],
+        ])
+            ->sole();
+
         $post->delete();
 
-        return \to_route('forum_topic', ['id' => $post->topic->id])
-            ->withSuccess(\trans('forum.delete-post-success'));
+        $latestPost = $topic->latestPost;
+        $isTopicDeleted = false;
+
+        if ($latestPost === null) {
+            $topic->delete();
+            $isTopicDeleted = true;
+        } else {
+            $latestPoster = $latestPost->user;
+            $topic->update([
+                'last_post_user_id'       => $latestPoster->id,
+                'last_post_user_username' => $latestPoster->username,
+                'num_post'                => $topic->posts()->count(),
+                'last_reply_at'           => $latestPost->created_at,
+            ]);
+        }
+
+        $forum = $topic->forum;
+        $lastRepliedTopic = $forum->lastRepliedTopic;
+        $latestPost = $lastRepliedTopic->latestPost;
+        $latestPoster = $latestPost->user;
+
+        $forum->update([
+            'num_post'                => $forum->posts()->count(),
+            'num_topic'               => $forum->topics()->count(),
+            'last_post_user_id'       => $latestPoster->id,
+            'last_post_user_username' => $latestPoster->username,
+            'last_topic_id'           => $lastRepliedTopic->id,
+            'last_topic_name'         => $lastRepliedTopic->name,
+            'updated_at'              => $latestPost->created_at,
+        ]);
+
+        if ($isTopicDeleted === true) {
+            return to_route('forums.show', ['id' => $forum->id])
+                ->withSuccess(trans('forum.delete-post-success'));
+        }
+
+        return to_route('topics.show', ['id' => $post->topic->id])
+            ->withSuccess(trans('forum.delete-post-success'));
     }
 }
